refactor(solvent): modularize clock, render, terminal; poll input before tick
Split clock, render, and terminal logic out of solvent lib.rs into dedicated
submodules to improve code organization. Add re-exports for the new public
functions. In the kernel scheduler, poll mouse and keyboard state before the
runtime tick to ensure the desktop remains responsive even on firmware/VM
configs without interrupt delivery. Also update timezone handling to use the
new clock module.
